+++ packages/runtime/runtime-utils/src/index.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-export * from "./componentHandleContextUtils";
+export * from "./dataStoreHandleContextUtils";
 export * from "./serializer";
 export * from "./summarizerNode";
 export * from "./summaryTracker";
