+++ packages/runtime/runtime-utils/src/test/requestParser.spec.ts
@@ -8,46 +8,46 @@ import { RequestParser } from "../requestParser";
 
 describe("RequestParser", () => {
     describe(".getPathParts", () => {
-        it("Parse Component Id", () => {
-            const url = "componentId";
+        it("Parse Data Store Id", () => {
+            const url = "dataStoreId";
             const pathParts = RequestParser.getPathParts(url);
             assert.equal(pathParts.length, 1);
-            assert.equal(pathParts[0], "componentId");
+            assert.equal(pathParts[0], "dataStoreId");
         });
-        it("Parse Component Id with /", () => {
-            const url = "/componentId/";
+        it("Parse Data Store Id with /", () => {
+            const url = "/dataStoreId/";
             const pathParts = RequestParser.getPathParts(url);
             assert.equal(pathParts.length, 1);
-            assert.equal(pathParts[0], "componentId");
+            assert.equal(pathParts[0], "dataStoreId");
         });
-        it("Parse Component Id with query", () => {
-            const url = "/componentId/?foo=bar";
+        it("Parse Data Store Id with query", () => {
+            const url = "/dataStoreId/?foo=bar";
             const pathParts = RequestParser.getPathParts(url);
             assert.equal(pathParts.length, 1);
-            assert.equal(pathParts[0], "componentId");
+            assert.equal(pathParts[0], "dataStoreId");
         });
-        it("Parse Component Id with sub route with query", () => {
-            const url = "/componentId//some/route?foo=bar";
+        it("Parse Data Store Id with sub route with query", () => {
+            const url = "/dataStoreId//some/route?foo=bar";
             const pathParts = RequestParser.getPathParts(url);
             assert.equal(pathParts.length, 3);
-            assert.equal(pathParts[0], "componentId");
+            assert.equal(pathParts[0], "dataStoreId");
             assert.equal(pathParts[1], "some");
             assert.equal(pathParts[2], "route");
         });
-        it("Parse encoded Component Id", () => {
-            const url = "component%20Id";
+        it("Parse encoded Data Store Id", () => {
+            const url = "data%20store%20Id";
             const pathParts = RequestParser.getPathParts(url);
             assert.equal(pathParts.length, 1);
-            assert.equal(pathParts[0], "component Id");
+            assert.equal(pathParts[0], "data store Id");
         });
     });
     describe(".createSubRequest", () => {
         let requestParser: RequestParser;
         beforeEach(() => {
-            requestParser = new RequestParser({ url: "/componentId//some/route/" });
+            requestParser = new RequestParser({ url: "/dataStoreId//some/route/" });
         });
         it("Create request from part 0", () => {
-            assert.equal(requestParser.createSubRequest(0)?.url, "componentId/some/route");
+            assert.equal(requestParser.createSubRequest(0)?.url, "dataStoreId/some/route");
         });
         it("Create request from part 1", () => {
             assert.equal(requestParser.createSubRequest(1)?.url, "some/route");
